+++ Database/Migrations/2020_11_25_155351_create_dormitory_room_table.php
@@ -16,10 +16,10 @@ class CreateDormitoryRoomTable extends Migration
         Schema::connection('mysql_dorm')->create('dormitory_room', function (Blueprint $table) {
             $table->increments('id');
             $table->string('roomnum')->default('')->comment('房间号');
-            $table->tinyInteger('buildtype')->comment('楼的类型');
+            $table->Integer('buildtype')->comment('楼的类型');
             $table->tinyInteger('floornum')->comment('所在楼层');
             $table->tinyInteger('bedsnum')->comment('床位数');
-            $table->tinyInteger('buildid')->comment('所在楼id');
+            $table->Integer('buildid')->comment('所在楼id');
             $table->timestamp('created_at')->nullable()->useCurrent()->comment('创建时间');
             $table->timestamp('updated_at')->nullable()->comment('更新时间');
         });
