+++ Database/Migrations/2020_11_25_155351_create_dormitory_users_building_table.php
@@ -16,7 +16,7 @@ class CreateDormitoryUsersBuildingTable extends Migration
         Schema::connection('mysql_dorm')->create('dormitory_users_building', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('idnum', 32)->index('username')->comment('教师工号');
-            $table->tinyInteger('buildid')->comment('楼宇id');
+            $table->Integer('buildid')->comment('楼宇id');
             $table->timestamp('created_at')->nullable()->useCurrent()->comment('创建时间');
             $table->softDeletes()->comment('删除时间');
             $table->timestamp('updated_at')->nullable()->comment('更新时间');
